add cpfEventMask on ProfileEditPerson && move eventMasks archive to utils
